apply Adrian's suggestions
